Move CRUD actions into add.js file and import them as modules, also fixed the index counting to start from one and fixed other requested chnages by the reviewer
