+++ src/active.js
@@ -1,10 +1,10 @@
 export const saveLocalStorage = (data) => {
   const todoItems = document.querySelectorAll('.todo-item');
   todoItems.forEach((item, idx) => {
-    item.id = idx;
+    item.id = idx + 1;
   });
   data.forEach((d, idx) => {
-    d.index = idx;
+    d.index = idx + 1;
   });
   localStorage.setItem('toDoStorage', JSON.stringify(data));
 };
